[feature/Refactor]feat: Merge AppleLoginTestViewModel into LoginViewModel
