+++ Owori/Owori/OworiApp.swift
@@ -18,13 +18,12 @@ struct OworiApp: App {
     var body: some Scene {
         
         WindowGroup {
-            AppleLoginTestView()
-//            LoginTestView()
-//                .onOpenURL { url in
-//                    if (AuthApi.isKakaoTalkLoginUrl(url)) {
-//                        _ = AuthController.handleOpenUrl(url: url)
-//                    }
-//                }
+            LoginTestView()
+                .onOpenURL { url in
+                    if (AuthApi.isKakaoTalkLoginUrl(url)) {
+                        _ = AuthController.handleOpenUrl(url: url)
+                    }
+                }
 
         }
     }
