+++ Owori/Owori/View/AppleLoginTestView.swift
@@ -9,16 +9,16 @@ import SwiftUI
 import AuthenticationServices
 
 struct AppleLoginTestView: View {
-    var appleLoginViewModel: AppleLoginTestViewModel = AppleLoginTestViewModel()
     var userViewModel: UserViewModel = UserViewModel()
+    var loginViewModel: LoginViewModel = LoginViewModel()
     var storyViewModel: StoryViewModel = StoryViewModel()
     var familyViewModel: FamilyViewModel = FamilyViewModel()
     
     var body: some View {
         VStack {
-            appleLoginViewModel.appleLoginButton()
+            loginViewModel.appleLoginButton()
             Button {
-                userViewModel.joinMember(socialToken: appleLoginViewModel.socialToken)
+                userViewModel.joinMember(socialToken: loginViewModel.socialToken)
             } label: {
                 Text("Create User")
             }
@@ -73,41 +73,3 @@ struct AppleLoginTestView: View {
     }
     
 }
-
-//struct AppleSigninButton : View{
-//    var body: some View{
-//        SignInWithAppleButton(
-//            onRequest: { request in
-//                request.requestedScopes = [.fullName, .email]
-//            },
-//            onCompletion: { result in
-//                switch result {
-//                case .success(let authResults):
-//                    print("Apple Login Successful")
-//                    switch authResults.credential{
-//                    case let appleIDCredential as ASAuthorizationAppleIDCredential:
-//                        // 계정 정보 가져오기
-//                        let UserIdentifier = appleIDCredential.user
-//                        let fullName = appleIDCredential.fullName
-//                        let name =  (fullName?.familyName ?? "") + (fullName?.givenName ?? "")
-//                        let email = appleIDCredential.email
-//                        let IdentityToken = String(data: appleIDCredential.identityToken!, encoding: .utf8)
-//                        let AuthorizationCode = String(data: appleIDCredential.authorizationCode!, encoding: .utf8)
-//
-//                        print("UserIdentifier: \(UserIdentifier)")
-//                        print("IdentityToken: \(IdentityToken)")
-//                        print("AuthorizationCode: \(AuthorizationCode)")
-//                    default:
-//                        break
-//                    }
-//                case .failure(let error):
-//                    print(error.localizedDescription)
-//                    print("error")
-//                }
-//            }
-//        )
-//        .frame(width : UIScreen.main.bounds.width * 0.9, height:50)
-//        .cornerRadius(5)
-//    }
-//}
-
